backend/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.8.1
astunparse==1.6.3
blinker==1.9.0
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
distlib==0.3.9
distro==1.9.0
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
elastic-transport==8.17.1
elasticsearch==8.17.2
Faker==37.1.0
filelock==3.16.1
firebase-admin==6.7.0
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.58.2
gast==0.6.0
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
h5py==3.12.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.10.0
joblib==1.4.2
keras==3.7.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml-dtypes==0.4.1
msgpack==1.1.0
namex==0.0.8
nltk==3.9.1
numpy==2.0.2
openai==1.106.1
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pipenv==2024.4.1
platformdirs==4.3.6
proto-plus==1.26.1
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
rsa==4.9
scapy==2.6.1
scikit-learn==1.6.1
scipy==1.14.1
seaborn==0.13.2
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tldextract==5.1.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.28.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
xgboost==3.0.0
